+++ Stripe/PaymentSheetViewController.swift
@@ -144,15 +144,12 @@ class PaymentSheetViewController: UIViewController {
                 return .setup
             }
         }()
-        
-        // TODO(porter) Remove hacky fix before appearance APIs are shipped
-        var appearanceCopy = configuration.appearance
-        appearanceCopy.color.primary = configuration.primaryButtonColor
-        
+
         let button = ConfirmButton(
             style: .stripe,
             callToAction: callToAction,
-            appearance: appearanceCopy,
+            appearance: configuration.appearance,
+            backgroundColor: configuration.primaryButtonColor,
             didTap: { [weak self] in
                 self?.didTapBuyButton()
             }
@@ -188,7 +185,7 @@ class PaymentSheetViewController: UIViewController {
         }
 
         super.init(nibName: nil, bundle: nil)
-        self.view.backgroundColor = configuration.appearance.color.background
+        self.view.backgroundColor = configuration.appearance.colors.background
 
         // Set the current elements theme
         ElementsUITheme.current = configuration.appearance.asElementsTheme
@@ -208,6 +205,7 @@ class PaymentSheetViewController: UIViewController {
         stackView.spacing = PaymentSheetUI.defaultPadding
         stackView.axis = .vertical
         stackView.bringSubviewToFront(headerLabel)
+        stackView.setCustomSpacing(32, after: paymentContainerView)
 
         // Hack: Payment container needs to extend to the edges, so we'll 'cancel out' the layout margins with negative padding
         paymentContainerView.directionalLayoutMargins = .insets(
@@ -227,8 +225,6 @@ class PaymentSheetViewController: UIViewController {
             stackView.bottomAnchor.constraint(
                 equalTo: view.bottomAnchor, constant: -PaymentSheetUI.defaultSheetMargins.bottom),
         ])
-
-        buyButton.tintColor = configuration.primaryButtonColor // TODO(porter): Read primary color for appearance
 
         updateUI(animated: false)
     }