+++ Stripe/STPImageLibrary.swift
@@ -110,9 +110,9 @@ public class STPImageLibrary: NSObject {
     public class func afterpayLogo(locale: Locale = Locale.current) -> UIImage {
         switch (locale.languageCode, locale.regionCode) {
             case ("en", "GB"):
-                return self.safeImageNamed("clearpay_mark", templateIfAvailable: false)
+                return self.safeImageNamed("clearpay_mark", templateIfAvailable: true)
             default:
-               return self.safeImageNamed("afterpay_mark", templateIfAvailable: false)
+               return self.safeImageNamed("afterpay_mark", templateIfAvailable: true)
         }
     }