+++ StripeIdentity/StripeIdentity/Source/VerificationClientSecret.swift
@@ -22,24 +22,20 @@ extension VerificationClientSecret {
      - returns: nil if the client secret is invalid
      */
     init?(string: String) {
-        // NOTE(mludowise): Setting `maxSplits` to `expectedComponentsCount`
-        // means that if there are too many underscores, the components will be
-        // equal to `expectedComponentsCount + 1`.
-        // This means strings like "vi__123_secret_456" will fail validation.
         let components = string
             .trimmingCharacters(in: .whitespacesAndNewlines)
             .split(separator: "_",
-                   maxSplits: VerificationClientSecret.expectedComponentsCount,
+                   maxSplits: VerificationClientSecret.expectedComponentsCount - 1,
                    omittingEmptySubsequences: false)
 
-        // Matching regex /^((vi|vs)_[0-9a-zA-Z]+)_secret_([0-9a-zA-Z]+)$/
-        guard components.count == VerificationClientSecret.expectedComponentsCount &&
+        // Matching regex /^((vi|vs)_[0-9a-zA-Z]+)_secret_(.+)$/
+        guard components.count >= VerificationClientSecret.expectedComponentsCount &&
                 (components[0] == "vi" || components[0] == "vs") &&
                 !components[1].isEmpty &&
                 (components[1].rangeOfCharacter(from: CharacterSet.alphanumerics.inverted) == nil) &&
                 components[2] == "secret" &&
-                !components[3].isEmpty &&
-                (components[3].rangeOfCharacter(from: CharacterSet.alphanumerics.inverted) == nil) else {
+                !components[3].isEmpty
+        else {
             return nil
         }